+++ src/cli/handlers.rs
@@ -43,6 +43,8 @@ mod task_cmd;
 mod up_cmd;
 mod update_cmd;
 mod url_cmd;
+#[cfg(unix)]
+mod v8_project_command;
 
 pub(crate) use about_cmd::handle_about;
 pub(crate) use app_create_cmd::{HandleAppCreateOptions, handle_app_create};
@@ -96,3 +98,5 @@ pub(crate) use task_cmd::{
 pub(crate) use up_cmd::{HandleUpOptions, handle_up};
 pub(crate) use update_cmd::{HandleUpdateOptions, handle_update};
 pub(crate) use url_cmd::handle_url;
+#[cfg(unix)]
+pub(crate) use v8_project_command::handle_v8_project_command;